refactor: Split dummy client files
